updated favoriteSlice.js file to fix favourite issue
